Refactor csv
